+++ src/richer_prompt/__init__.py
@@ -1,7 +1,7 @@
 from richer_prompt import keys
 from richer_prompt._version import get_version_dict
 from richer_prompt.choices import Choice
-from richer_prompt.prompt import MultiSelect, Select, Tabs
+from richer_prompt.prompt import Form, MultiSelect, Select, Tabs
 from richer_prompt.session import NotInteractiveError
 
 __all__ = [
@@ -9,11 +9,11 @@ __all__ = [
     "Select",
     "MultiSelect",
     "Tabs",
+    "Form",
     "Choice",
     "NotInteractiveError",
     "keys",
 ]
 
 __version__ = get_version_dict()["version"]
-
 del get_version_dict